Refactor parameter slightly and fix #871
