+++ policyengine-client/src/policyengine/pages/policy/parameter/index.js
@@ -1,0 +1,2 @@
+import Parameter from "./parameter";
+export default Parameter;
